Add Session

+++ application/controllers/admin/Contacts.php
@@ -1,9 +1,4 @@
-<?php
-	/*
-    @Copyright Indra Rukmana
-    @Class Name : Contacts
-	*/
-defined('BASEPATH') OR exit('No direct script access allowed');
+<?php defined('BASEPATH') OR exit('No direct script access allowed');
 
 class Contacts extends CI_Controller {
 
@@ -15,23 +10,31 @@ class Contacts extends CI_Controller {
 	
 	// Inbox Contacts
 	public function inbox() {
+		if($this->session->userdata('logged_in')!="" && $this->session->userdata('status')=="administrator") {
 
-		$contacts = $this->mContacts->listContacts();
-		
-		$data = array(	'judul_lengkap'	=> $this->config->item('nama_aplikasi_full'),
-						'judul_pendek'	=> $this->config->item('nama_aplikasi_pendek'),
-						'instansi'		=> $this->config->item('nama_instansi'),
-						'credit'		=> $this->config->item('credit_aplikasi'),
-						'contacts'	=> $contacts,
-						'isi'		=> 'admin/message/list');
-		$this->load->view('admin/layout/wrapper',$data);
+			$contacts = $this->mContacts->listContacts();
+			
+			$data = array(	'judul_lengkap'	=> $this->config->item('nama_aplikasi_full'),
+							'judul_pendek'	=> $this->config->item('nama_aplikasi_pendek'),
+							'instansi'		=> $this->config->item('nama_instansi'),
+							'credit'		=> $this->config->item('credit_aplikasi'),
+							'contacts'	=> $contacts,
+							'isi'		=> 'admin/message/list');
+			$this->load->view('admin/layout/wrapper',$data);
+		}else{
+			redirect('login');
+		}
 	}
 
 	// Delete Message
 	public function delete($message_id) {
-		$data = array('message_id'	=> $message_id);
-		$this->mContacts->deleteMessage($data);		
-		$this->session->set_flashdata('sukses','Success');
-		redirect(base_url('admin/contacts/inbox'));
+		if($this->session->userdata('logged_in')!="" && $this->session->userdata('status')=="administrator") {
+			$data = array('message_id'	=> $message_id);
+			$this->mContacts->deleteMessage($data);		
+			$this->session->set_flashdata('sukses','Success');
+			redirect(base_url('admin/contacts/inbox'));
+		}else{
+			redirect('login');
+		}
 	}	
 }
